agregando security

+++ src/main/java/com/prestamos/repository/RolRepository.java
@@ -4,10 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.prestamos.model.Rol;
 
-import java.util.List;
 
-
-public interface RolRepository extends JpaRepository<Rol,Long> {
+public interface RolRepository extends JpaRepository<Rol,Integer>  {
 
 	
 }
