+++ src/main/java/com/prestamos/repository/ZonaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.prestamos.model.Zona;
 
-public interface ZonaRepository extends JpaRepository<Zona,Long> {
+public interface ZonaRepository extends JpaRepository<Zona,Integer> {
 }
